+++ src/console/module/console.module.ts
@@ -1,0 +1,10 @@
+import { Module } from '@nestjs/common';
+import { ConsoleInterceptor } from '../ConsoleInterceptor'
+import { ConsoleController } from '../controller/console.controller'
+
+@Module({
+  providers: [ConsoleInterceptor],
+  exports: [ConsoleInterceptor],
+  controllers: [ConsoleController]
+})
+export class ConsoleModule {}